refactor(fsutil): promote replaceFile to fsutil.ReplaceFile (poster-write P3, slice 1)
The replace-with-rename primitive (POSIX rename; Windows MoveFileEx with
MOVEFILE_REPLACE_EXISTING, wrapped ErrReplaceUnsupported on virtual FS)
moves from internal/downloader to internal/fsutil so the revert-ledger
backup/replace flow can share it. Downloader call sites adopt it; tests
travel with the code unchanged apart from export naming.
