--- internal/downloader/replacefile.go
@@ -1,1 +1,0 @@
-package downloader
