+++ internal/fsutil/replacefile.go
@@ -1,0 +1,1 @@
+package fsutil
