+++ internal/fsutil/replacefile_test.go
@@ -1,14 +1,30 @@
-package downloader
+package fsutil
 
 import (
+	"errors"
 	"os"
 	"sync/atomic"
 	"testing"
 
 	"github.com/spf13/afero"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
+
+// rejectExistingRenameFS mimics a Windows-style rename that refuses existing
+// destinations (ReplaceFile on virtual FS must surface the failure and keep
+// the pre-existing bytes).
+type rejectExistingRenameFS struct {
+	afero.Fs
+}
+
+func (f rejectExistingRenameFS) Rename(oldname, newname string) error {
+	if _, err := f.Fs.Stat(newname); err == nil {
+		return errors.New("replace existing destination rejected")
+	}
+	return f.Fs.Rename(oldname, newname)
+}
 
 type recordingRenameFS struct {
 	afero.Fs
@@ -25,7 +41,7 @@ func TestReplaceFile_ReplacesExistingOnVirtualFS(t *testing.T) {
 	require.NoError(t, afero.WriteFile(fs, "/output/source.tmp", []byte("new"), 0644))
 	require.NoError(t, afero.WriteFile(fs, "/output/destination.jpg", []byte("old"), 0644))
 
-	require.NoError(t, replaceFile(fs, "/output/source.tmp", "/output/destination.jpg"))
+	require.NoError(t, ReplaceFile(fs, "/output/source.tmp", "/output/destination.jpg"))
 	got, err := afero.ReadFile(fs, "/output/destination.jpg")
 	require.NoError(t, err)
 	assert.Equal(t, []byte("new"), got)
@@ -39,7 +55,7 @@ func TestReplaceFile_DispatchesToWrappedBackend(t *testing.T) {
 	require.NoError(t, afero.WriteFile(fs, "/output/source.tmp", []byte("new"), 0644))
 	require.NoError(t, afero.WriteFile(fs, "/output/destination.jpg", []byte("old"), 0644))
 
-	require.NoError(t, replaceFile(fs, "/output/source.tmp", "/output/destination.jpg"))
+	require.NoError(t, ReplaceFile(fs, "/output/source.tmp", "/output/destination.jpg"))
 	assert.Equal(t, int32(1), fs.calls.Load())
 	got, err := afero.ReadFile(base, "/output/destination.jpg")
 	require.NoError(t, err)
@@ -53,7 +69,7 @@ func TestReplaceFile_FailurePreservesDestination(t *testing.T) {
 	old := []byte("old")
 	require.NoError(t, afero.WriteFile(base, "/output/destination.jpg", old, 0644))
 
-	err := replaceFile(fs, "/output/source.tmp", "/output/destination.jpg")
+	err := ReplaceFile(fs, "/output/source.tmp", "/output/destination.jpg")
 	require.Error(t, err)
 	got, readErr := afero.ReadFile(base, "/output/destination.jpg")
 	require.NoError(t, readErr)
